Imported Upstream version 2.0.5

+++ src/plugins/select/bluegene/plugin/bg_job_place.c
@@ -97,7 +97,8 @@ static int _check_for_booted_overlapping_blocks(
 	bool test_only);
 static int _dynamically_request(List block_list, int *blocks_added,
 				ba_request_t *request,
-				char *user_req_nodes);
+				char *user_req_nodes,
+				bool test_only);
 static int _find_best_block_match(List block_list, int *blocks_added,
 				  struct job_record* job_ptr,
 				  bitstr_t* slurm_block_bitmap,
@@ -668,6 +669,10 @@ static int _check_for_booted_overlapping_blocks(
 								bg_record);
 					}
 
+					debug("Removing unusable block %s "
+					      "from the system.",
+					      bg_record->bg_block_id);
+					
 					if(!found_record) {
 						debug("This record %s wasn't "
 						      "found in the "
@@ -678,10 +683,6 @@ static int _check_for_booted_overlapping_blocks(
 						found_record = bg_record;
 					} else
 						destroy_bg_record(bg_record);
-					
-					debug("removing the block %s"
-					      "from the system",
-					      bg_record->bg_block_id);
 					
 					list_push(temp_list, found_record);
 					free_block_list(temp_list);
@@ -708,7 +709,8 @@ static int _check_for_booted_overlapping_blocks(
 
 static int _dynamically_request(List block_list, int *blocks_added,
 				ba_request_t *request,
-				char *user_req_nodes)
+				char *user_req_nodes,
+				bool test_only)
 {
 	List list_of_lists = NULL;
 	List temp_list = NULL;
@@ -755,7 +757,10 @@ static int _dynamically_request(List block_list, int *blocks_added,
 			while((bg_record = list_pop(new_blocks))) {
 				if(block_exist_in_list(block_list, bg_record))
 					destroy_bg_record(bg_record);
-				else {
+				else if(test_only) {
+					list_append(block_list, bg_record);
+					(*blocks_added) = 1;
+				} else {
 					if(job_block_test_list 
 					   == bg_lists->job_running) {
 						if(configure_block(bg_record)
@@ -774,18 +779,18 @@ static int _dynamically_request(List block_list, int *blocks_added,
 					list_append(block_list, bg_record);
 					print_bg_record(bg_record);
 					(*blocks_added) = 1;
-				}
+				} 	
 			}
 			list_destroy(new_blocks);
 			if(!*blocks_added) {
-				memcpy(request->geometry, start_geo,      
+				memcpy(request->geometry, start_geo,
 				       sizeof(int)*BA_SYSTEM_DIMENSIONS); 
 				rc = SLURM_ERROR;
 				continue;
 			}
 			list_sort(block_list,
 				  (ListCmpF)_bg_record_sort_aval_dec);
-	
+			
 			rc = SLURM_SUCCESS;
 			break;
 		} else if (errno == ESLURM_INTERCONNECT_FAILURE) {
@@ -907,57 +912,8 @@ static int _find_best_block_match(List block_list,
 		}
 		if(!req_nodes)
 			req_nodes = min_nodes;
-	}
-	if (target_size == 0) {	/* no geometry specified */
-		if(job_ptr->details->req_nodes 
-		   && !start_req) {
-			bg_record_t *tmp_record = NULL;
-			char *tmp_nodes= job_ptr->details->req_nodes;
-			int len = strlen(tmp_nodes);
-			
-			i = 0;
-			while(i<len 
-			      && tmp_nodes[i] != '[' 
-			      && (tmp_nodes[i] < '0' || tmp_nodes[i] > 'Z'
-				  || (tmp_nodes[i] > '9'
-				      && tmp_nodes[i] < 'A')))
-				i++;
-			
-			if(i<len) {
-				len -= i;
-				tmp_record = xmalloc(sizeof(bg_record_t));
-				tmp_record->bg_block_list =
-					list_create(destroy_ba_node);
-				
-				len += strlen(bg_conf->slurm_node_prefix)+1;
-				tmp_record->nodes = xmalloc(len);
-				
-				snprintf(tmp_record->nodes,
-					 len,
-					 "%s%s", 
-					 bg_conf->slurm_node_prefix, 
-					 tmp_nodes+i);
-				
-			
-				process_nodes(tmp_record, false);
-				for(i=0; i<BA_SYSTEM_DIMENSIONS; i++) {
-					req_geometry[i] = tmp_record->geo[i];
-					start[i] = tmp_record->start[i];
-				}
-				destroy_bg_record(tmp_record);
-				select_g_set_jobinfo(job_ptr->select_jobinfo,
-						     SELECT_DATA_GEOMETRY, 
-						     &req_geometry);
-				select_g_set_jobinfo(job_ptr->select_jobinfo,
-						     SELECT_DATA_START, 
-						     &start);
-				start_req = 1;
-			}  else 
-				error("BPs=%s is in a weird format", 
-				      tmp_nodes); 
-		} else {
-			req_geometry[X] = (uint16_t)NO_VAL;
-		}
+	} else {
+		req_geometry[X] = (uint16_t)NO_VAL;
 		target_size = min_nodes;
 	}
 	
@@ -1090,7 +1046,8 @@ static int _find_best_block_match(List block_list,
 		
 		if((rc = _dynamically_request(block_list, blocks_added,
 					      &request, 
-					      job_ptr->details->req_nodes))
+					      job_ptr->details->req_nodes,
+					      test_only))
 		   == SLURM_SUCCESS) {
 			create_try = 1;
 			continue;
@@ -1147,8 +1104,17 @@ static int _find_best_block_match(List block_list,
 					destroy_bg_record(bg_record);
 					if(errno == ESLURM_INTERCONNECT_FAILURE
 					   || !list_count(job_list)) {
-						error("this job will never "
-						      "run on this system");
+						char *nodes;
+						if (slurmctld_conf.
+						    slurmctld_debug < 5)
+							break;
+						nodes = bitmap2node_name(
+							slurm_block_bitmap);
+						debug("job %u not "
+						      "runable on %s",
+						      job_ptr->job_id,
+						      nodes);
+						xfree(nodes);
 						break;
 					}
 					continue;
@@ -1238,6 +1204,11 @@ static int _sync_block_lists(List full_list, List incomp_list)
 	itr = list_iterator_create(full_list);
 	itr2 = list_iterator_create(incomp_list);
 	while((new_record = list_next(itr))) {
+		/* Make sure we aren't adding any block that doesn't
+		   have a block_id.
+		*/
+		if(!new_record->bg_block_id)
+			continue;
 		while((bg_record = list_next(itr2))) {
 			if(bit_equal(bg_record->bitmap, new_record->bitmap)
 			   && bit_equal(bg_record->ionode_bitmap,
@@ -1246,10 +1217,9 @@ static int _sync_block_lists(List full_list, List incomp_list)
 		} 
 
 		if(!bg_record) {
-			bg_record = xmalloc(sizeof(bg_record_t));
-			copy_bg_record(new_record, bg_record);
-			debug4("adding %s", bg_record->bg_block_id);
-			list_append(incomp_list, bg_record);
+			list_remove(itr);
+			debug4("adding %s", new_record->bg_block_id);
+			list_append(incomp_list, new_record);
 			count++;
 		} 
 		list_iterator_reset(itr2);
@@ -1385,8 +1355,8 @@ extern int submit_job(struct job_record *job_ptr, bitstr_t *slurm_block_bitmap,
 	}
 	select_g_sprint_jobinfo(job_ptr->select_jobinfo, buf, sizeof(buf), 
 				SELECT_PRINT_MIXED);
-	debug("bluegene:submit_job: %s nodes=%u-%u-%u", 
-	      buf, min_nodes, req_nodes, max_nodes);
+	debug("bluegene:submit_job: %d %s nodes=%u-%u-%u", 
+	      mode, buf, min_nodes, req_nodes, max_nodes);
 	select_g_sprint_jobinfo(job_ptr->select_jobinfo, buf, sizeof(buf), 
 				SELECT_PRINT_BLRTS_IMAGE);
 #ifdef HAVE_BGL
@@ -1450,29 +1420,29 @@ extern int submit_job(struct job_record *job_ptr, bitstr_t *slurm_block_bitmap,
 					     bg_record->ionodes);
 			
 			if(!bg_record->bg_block_id) {
-				uint16_t geo[BA_SYSTEM_DIMENSIONS];
-				
 				debug2("%d can start unassigned job %u at "
 				       "%u on %s",
 				       test_only, job_ptr->job_id, starttime,
 				       bg_record->nodes);
 				select_g_set_jobinfo(job_ptr->select_jobinfo,
-					     SELECT_DATA_BLOCK_ID,
-					     "unassigned");
-
+						     SELECT_DATA_BLOCK_ID,
+						     "unassigned");
+				
 				min_nodes = bg_record->node_cnt;
 				select_g_set_jobinfo(job_ptr->select_jobinfo,
-					     SELECT_DATA_NODE_CNT,
-					     &min_nodes);
-				memset(geo, 0, 
-				       sizeof(uint16_t) * BA_SYSTEM_DIMENSIONS);
-				select_g_set_jobinfo(job_ptr->select_jobinfo,
-						     SELECT_DATA_GEOMETRY, 
-						     &geo);
+						     SELECT_DATA_NODE_CNT,
+						     &min_nodes);
 				/* This is a fake record so we need to
 				 * destroy it after we get the info from
-				 * it */
-				destroy_bg_record(bg_record);
+				 * it.  if it was just testing then
+				 * we added this record to the
+				 * block_list.  If this is the case
+				 * it will be set below, but set
+				 * blocks_added to 0 since we don't
+				 * want to sync this with the list. */
+				if(!blocks_added)
+					destroy_bg_record(bg_record);
+				blocks_added = 0;
 			} else {
 				if((bg_record->ionodes)
 				   && (job_ptr->part_ptr->max_share <= 1))
@@ -1490,9 +1460,6 @@ extern int submit_job(struct job_record *job_ptr, bitstr_t *slurm_block_bitmap,
 				select_g_set_jobinfo(job_ptr->select_jobinfo,
 						     SELECT_DATA_NODE_CNT, 
 						     &bg_record->node_cnt);
-				select_g_set_jobinfo(job_ptr->select_jobinfo,
-						     SELECT_DATA_GEOMETRY, 
-						     &bg_record->geo);
 
 				/* tmp16 = bg_record->conn_type; */
 /* 				select_g_set_jobinfo(job_ptr->select_jobinfo, */
